Count units in adapter
